Create rock, paper, scissors game in JavaScript
Did the entirety of this project without commiting. Had been a
few days without doing much coding and kind of got lost in
writing it out and just forgot to commit. On the bright side, that
happened because I was having a lot of fun problem solving on this
project. I will likely come back to this one a few times after this
second commit to touch just a few things up that aren't specified in
the project, such as looping for invalid input once I learn it.